initial refactoring stepts and output pipeline
