+++ misc/print_stats.py
@@ -1,0 +1,69 @@
+# print
+def print_batch_stats(model, epoch, batch_idx, train_loader, losses, params):
+    '''
+    prints statistics about the recently seen batches
+    '''
+    print('Epoch: {} of {}'.format(epoch, params.n_epochs))
+    print('Batch: {} of {}'.format(batch_idx, len(train_loader)))
+    print('Loss past {} batches: Localization {} Classification {}'.format(params.train_stats_step,
+                                                                           losses[0] / params.train_stats_step, losses[1] / params.train_stats_step))
+
+    mean_grads, max_grads, mean_weights, max_weights = gradient_weight_check(model)
+
+    print('Mean and max gradients over whole network: ', mean_grads, max_grads)
+    print('Mean and max weights over whole network: ', mean_weights, max_weights)
+
+
+# print stats
+def gradient_weight_check(model):
+    '''
+    will pring mean abs value of gradients and weights during training to check for stability
+    '''
+    avg_grads, max_grads = [], []
+    avg_weigths, max_weigths = [], []
+
+    # try to understand comp graph better for why inter variables don't have grad retained and what this means for this stat
+    for n, p in model.named_parameters():
+        if (p.requires_grad) and (type(p.grad) != type(None)):
+            avg_grads.append(p.grad.abs().mean())
+            max_grads.append(p.grad.abs().max())
+            avg_weigths.append(p.abs().mean())
+            max_weigths.append(p.abs().max())
+
+    avg_grads, max_grads = torch.FloatTensor(avg_grads), torch.FloatTensor(max_grads)
+    avg_weigths, max_weigths = torch.FloatTensor(avg_weigths), torch.FloatTensor(max_weigths)
+
+    return torch.mean(avg_grads), torch.mean(max_grads), torch.mean(avg_weigths), torch.mean(max_weigths)
+
+# print stats
+def plot_grad_flow(model):
+    # taken from Roshan Rane answer on pytorch forums
+    '''Plots the gradients flowing through different layers in the net during training.
+    Can be used for checking for possible gradient vanishing / exploding problems.
+
+    Usage: Plug this function in Trainer class after loss.backwards() as
+    "plot_grad_flow(self.model.named_parameters())" to visualize the gradient flow
+
+    - will want to extend this to write ave_grads and max_grads to a simple csv file and plot progressions after training
+    '''
+    ave_grads = []
+    max_grads = []
+    layers = []
+    for n, p in model.named_parameters():
+        if(p.requires_grad) and ("bias" not in n):
+            layers.append(n)
+            ave_grads.append(p.grad.abs().mean())
+            max_grads.append(p.grad.abs().max())
+    plt.bar(np.arange(len(max_grads)), max_grads, alpha=0.1, lw=1, color="c")
+    plt.bar(np.arange(len(max_grads)), ave_grads, alpha=0.1, lw=1, color="b")
+    plt.hlines(0, 0, len(ave_grads)+1, lw=2, color="k")
+    plt.xticks(range(0, len(ave_grads), 1), layers, rotation="vertical")
+    plt.xlim(left=0, right=len(ave_grads))
+    plt.ylim(bottom=-0.001, top=0.02)  # zoom in on the lower gradient regions
+    plt.xlabel("Layers")
+    plt.ylabel("average gradient")
+    plt.title("Gradient flow")
+    plt.grid(True)
+    plt.legend([Line2D([0], [0], color="c", lw=4),
+                Line2D([0], [0], color="b", lw=4),
+                Line2D([0], [0], color="k", lw=4)], ['max-gradient', 'mean-gradient', 'zero-gradient'])
